fixed issue

+++ ext.cpp
@@ -31,7 +31,7 @@ pybind11::dict Registrations() {
   return dict;
 }
 
-auto BuildFwdDescriptor(int image_height, int image_width, int degree, int P, double tan_fovx, double tan_fovy) {
+auto BuildFwdDescriptor(int image_height, int image_width, int degree, int P, float tan_fovx, float tan_fovy) {
   FwdDescriptor d;
   d.image_height = image_height;
   d.image_width = image_width;
